requirements.txt
# App atuliazado em 07/2022
# Criado em VS Code com configurações para o mesmo
# Sistema operacional Windowns

aniso8601==9.0.1
astroid==2.11.7
autopep8==1.6.0
certifi==2022.6.15       
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
flake8==4.0.1
Flask==2.1.2
Flask-JWT-Extended==4.4.2
Flask-RESTful==0.3.9     
Flask-SQLAlchemy==2.5.1  
greenlet==1.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
platformdirs==2.5.2
protobuf==4.21.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pylint==2.14.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
types-requests==2.28.1
types-urllib3==1.26.16
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.0.0
wrapt==1.14.1